add data, pics, and C development part

+++ C_development/source/pure_c_lbm.c
@@ -14,27 +14,27 @@ const int ex[Q] = {0, 1, 0, -1, 0, 1, -1, -1, 1};
 const int ey[Q] = {0, 0, 1, 0, -1, 1, 1, -1, -1};
 
 // fi
-float fi[xdim + 2][ydim + 2][Q];
-float feq[xdim + 2][ydim + 2][Q];
-float fi_star[xdim + 2][ydim + 2][Q] = {0.0};
+double fi[xdim + 2][ydim + 2][Q];
+double feq[xdim + 2][ydim + 2][Q];
+double fi_star[xdim + 2][ydim + 2][Q] = {0.0};
 
 // Wi
-const float wi[Q] = {4.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0};
+const double wi[Q] = {4.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 9.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0, 1.0 / 36.0};
 
 // macro fluid density rho
-float rho[xdim + 2][ydim + 2];
+double rho[xdim + 2][ydim + 2];
 
-float u_x[xdim + 2][ydim + 2] = {0.001};
-float u_y[xdim + 2][ydim + 2] = {1.0};
+double u_x[xdim + 2][ydim + 2] = {0.001};
+double u_y[xdim + 2][ydim + 2] = {1.0};
 
 // parameters for check
-float mass = 0.0;
+double mass = 0.0;
 
 // parameter for the checking problem
-float U_0 = 0.01;
-float K = 30.0;
-float delta = 0.05;
-float cVel = 1 / sqrt(3); // c_s
+double U_0 = 0.01;
+double K = 30.0;
+double delta = 0.05;
+double cVel = 1 / sqrt(3); // c_s
 
 /*
  * Init the velocity u, v = (u_x, u_y)
@@ -48,15 +48,15 @@ float cVel = 1 / sqrt(3); // c_s
  * 
  * v = delta sin( 2pi *x / xdim )
 */
-void initRho_N_U(float (*u_x)[ydim + 2], float (*u_y)[ydim + 2], float (*rho)[ydim + 2])
+void initRho_N_U(double (*u_x)[ydim + 2], double (*u_y)[ydim + 2], double (*rho)[ydim + 2])
 {
     for (int i = 1; i < xdim + 1; i++)
     {
         for (int j = 1; j < ydim + 1; j++)
         {
-            float x_temp, y_temp;
-            x_temp = 1.0 * (float)(i - 1) / (float)xdim;
-            y_temp = 1.0 * (float)(j - 1) / (float)ydim; // scale the i and j to make 0 < i,j < 1
+            double x_temp, y_temp;
+            x_temp = 1.0 * (double)(i - 1) / (double)xdim;
+            y_temp = 1.0 * (double)(j - 1) / (double)ydim; // scale the i and j to make 0 < i,j < 1
 
             if (y_temp <= 0.5)
             {
@@ -73,9 +73,9 @@ void initRho_N_U(float (*u_x)[ydim + 2], float (*u_y)[ydim + 2], float (*rho)[yd
     }
 }
 
-float findBig(float (*u)[ydim + 2])
+double findBig(double (*u)[ydim + 2])
 {
-    float big = 0.0;
+    double big = 0.0;
     for (int i = 1; i < xdim + 1; i++)
     {
         for (int j = 1; j < ydim + 1; j++)
@@ -86,9 +86,9 @@ float findBig(float (*u)[ydim + 2])
     }
     return big;
 }
-float findSmall(float (*u)[ydim + 2])
+double findSmall(double (*u)[ydim + 2])
 {
-    float small = 0.0;
+    double small = 0.0;
     for (int i = 1; i < xdim + 1; i++)
     {
         for (int j = 1; j < ydim + 1; j++)
@@ -123,12 +123,12 @@ int getDelta(int a, int b)
 /*
  * init the distribution function fi 
 */
-void initFi(float (*fi)[ydim + 2][Q], float (*rho)[ydim + 2])
+void initFi(double (*fi)[ydim + 2][Q], double (*rho)[ydim + 2])
 {
 
     // for check problem
-    float A = 1 / (cVel * cVel);
-    float B = 1 / 2 * cVel * cVel * cVel * cVel;
+    double A = 1 / (cVel * cVel);
+    double B = 1 / 2 * cVel * cVel * cVel * cVel;
     for (int i = 0; i <= xdim + 1; i++)
     {
         for (int j = 0; j <= ydim + 1; j++)
@@ -178,9 +178,9 @@ void printfeq()
  *  Use computeFeq() instead !!
  */
 
-void initFeq(float (*rho)[ydim + 2], float (*feq)[ydim + 2][Q])
+void initFeq(double (*rho)[ydim + 2], double (*feq)[ydim + 2][Q])
 {
-    float c_s = 1 / sqrt(3); // the speed of sound in the lattice
+    double c_s = 1 / sqrt(3); // the speed of sound in the lattice
     for (int i = 0; i <= xdim + 1; i++)
     {
         for (int j = 0; j <= ydim + 1; j++)
@@ -195,7 +195,7 @@ void initFeq(float (*rho)[ydim + 2], float (*feq)[ydim + 2][Q])
 }
 
 // move fi --> fi_star in the direction of ei
-void streamStep(float (*fi_star)[ydim + 2][Q], float (*fi)[ydim + 2][Q])
+void streamStep(double (*fi_star)[ydim + 2][Q], double (*fi)[ydim + 2][Q])
 {
 
     // periodic condition
@@ -251,7 +251,7 @@ void printfi_star()
  * compute the macroscopic density rho and velocity u, u = (ux, uy) for every lattice
 */
 
-void computeRho_N_U(float (*rho)[ydim + 2], float (*u_x)[ydim + 2], float (*u_y)[ydim + 2], float (*fi)[ydim + 2][Q])
+void computeRho_N_U(double (*rho)[ydim + 2], double (*u_x)[ydim + 2], double (*u_y)[ydim + 2], double (*fi)[ydim + 2][Q])
 {
     for (int i = 1; i < xdim + 1; i++)
     {
@@ -276,14 +276,14 @@ void computeRho_N_U(float (*rho)[ydim + 2], float (*u_x)[ydim + 2], float (*u_y)
 * Compute the equilibrium distribution function f_a^eq for every lattice
 * int ex[Q] = {0, 1, 0, -1, 0, 1, -1, -1, 1};
 * int ey[Q] = {0, 0, 1, 0, -1, 1, 1, -1, -1};
-* float **u_x, float **u_y
+* double **u_x, double **u_y
 */
-void computeFeq(float (*rho)[ydim + 2], float (*u_x)[ydim + 2], float (*u_y)[ydim + 2], float (*feq)[ydim + 2][Q])
+void computeFeq(double (*rho)[ydim + 2], double (*u_x)[ydim + 2], double (*u_y)[ydim + 2], double (*feq)[ydim + 2][Q])
 {
-    float f1 = 3.0;
-    float f2 = 9.0 / 2.0;
-    float f3 = 3.0 / 2.0;
-    float uuDot, euDot, si;
+    double f1 = 3.0;
+    double f2 = 9.0 / 2.0;
+    double f3 = 3.0 / 2.0;
+    double uuDot, euDot, si;
 
     for (int i = 1; i < xdim + 1; i++)
     {
@@ -305,7 +305,7 @@ void computeFeq(float (*rho)[ydim + 2], float (*u_x)[ydim + 2], float (*u_y)[ydi
 * Collide Step : calculate the update distribution function fi = fi_Star - (fi_star - feq) / tau
 */
 
-void collideStep(float (*fi)[ydim + 2][Q], float (*feq)[ydim + 2][Q], float (*fi_star)[ydim + 2][Q], float tau)
+void collideStep(double (*fi)[ydim + 2][Q], double (*feq)[ydim + 2][Q], double (*fi_star)[ydim + 2][Q], double tau)
 {
     for (int i = 1; i < xdim + 1; i++)
     {
@@ -319,7 +319,7 @@ void collideStep(float (*fi)[ydim + 2][Q], float (*feq)[ydim + 2][Q], float (*fi
     }
 }
 
-void printRho(float (*rho)[ydim + 2])
+void printRho(double (*rho)[ydim + 2])
 {
     printf("\nthe rho metics is: \n");
     for (int i = 1; i < xdim + 1; i++)
@@ -333,7 +333,7 @@ void printRho(float (*rho)[ydim + 2])
     printf("\n");
 }
 
-void calWholeDensity(float (*fi)[ydim + 2][Q], float mass)
+void calWholeDensity(double (*fi)[ydim + 2][Q], double mass)
 {
     mass = 0.0;
     for (int i = 1; i < xdim + 1; i++)
@@ -349,9 +349,9 @@ void calWholeDensity(float (*fi)[ydim + 2][Q], float mass)
     printf("The total mass is %f \n", mass);
 }
 
-void calWholeMomentum(float (*rho)[ydim + 2], float (*u_x)[ydim + 2], float (*u_y)[ydim + 2])
+void calWholeMomentum(double (*rho)[ydim + 2], double (*u_x)[ydim + 2], double (*u_y)[ydim + 2])
 {
-    float momentum[2] = {0.0};
+    double momentum[2] = {0.0};
     for (int i = 1; i < xdim + 1; i++)
     {
         for (int j = 1; j < ydim + 1; j++)
@@ -363,7 +363,7 @@ void calWholeMomentum(float (*rho)[ydim + 2], float (*u_x)[ydim + 2], float (*u_
     printf("the Momentum in X is %f, in y is %f \n", momentum[0], momentum[1]);
 }
 
-void checkCorrectness(int iteration, float (*rho)[ydim + 2], float (*u_x)[ydim + 2], float (*u_y)[ydim + 2], float (*fi)[ydim + 2][Q], float mass)
+void checkCorrectness(int iteration, double (*rho)[ydim + 2], double (*u_x)[ydim + 2], double (*u_y)[ydim + 2], double (*fi)[ydim + 2][Q], double mass)
 {
     if (iteration % 1000 == 0)
     {
@@ -375,7 +375,7 @@ void checkCorrectness(int iteration, float (*rho)[ydim + 2], float (*u_x)[ydim +
 /// make the calcualation
 ///
 
-void calVorticity(float (*u_x)[ydim + 2], float (*u_y)[ydim + 2])
+void calVorticity(double (*u_x)[ydim + 2], double (*u_y)[ydim + 2])
 {
     FILE *fp;
     fp = fopen("Vorticity.csv", "w");
@@ -383,7 +383,7 @@ void calVorticity(float (*u_x)[ydim + 2], float (*u_y)[ydim + 2])
     {
         for (int j = 1; j < ydim + 1; j++)
         {
-            float voriticity = 0.5 * (u_y[i + 1][j] - u_y[i - 1][j]) + 0.5 * (u_x[i][j + 1] - u_x[i][j - 1]);
+            double voriticity = 0.5 * (u_y[i + 1][j] - u_y[i - 1][j]) + 0.5 * (u_x[i][j + 1] - u_x[i][j - 1]);
             fprintf(fp, "%.7f,", voriticity);
         }
         fprintf(fp, "\n");
@@ -404,7 +404,7 @@ void printUxAndUy()
     }
 }
 
-void showTheParameters(int maxiteration, float nu, float tau)
+void showTheParameters(int maxiteration, double nu, double tau)
 {
     printf("The size of the simulation is %d * %d,\nThe max iteration is %d\nU_0 = 0.01\nK = 30.0\ndelta = 0.05\n", xdim, ydim, maxiteration);
     printf("Simulation Start:\n");
@@ -413,11 +413,11 @@ void showTheParameters(int maxiteration, float nu, float tau)
 int main(int argc, char const *argv[])
 {
 
-    float big, small;
+    double big, small;
     int maxiterations = 12000;
     int iterations = 0;
-    float nu = 0.000128;
-    float tau = (1 + nu * 6) / 2;
+    double nu = 0.000128;
+    double tau = (1 + nu * 6) / 2;
     initRho_N_U(u_x, u_y, rho);
     initFi(fi, rho);
     computeFeq(rho, u_x, u_y, feq);
